Fixed rounding error for damage method

+++ Heroes/Hero.cs
@@ -27,7 +27,7 @@ namespace RPG_Heroes.Heroes
         public abstract void LevelUp();
         public abstract void Equip(Weapon weapon);
         public abstract void Equip(Armor armor);
-        public abstract int Damage();
+        public abstract double Damage();
         public abstract string Display();
 
         public HeroAttribute TotalAttributes()
